feat/water-2636 --documents service code refactored and tests fixed
